SDR-119: refactor: Authority 를 Domain 으로 이동
