--- be/src/main/java/com/jjikmuk/sikdorak/user/auth/controller/Authority.java
@@ -1,7 +1,0 @@
-package com.jjikmuk.sikdorak.user.auth.controller;
-
-public enum Authority {
-
-	USER, ANONYMOUS
-
-}
