+++ be/src/main/java/com/jjikmuk/sikdorak/user/user/domain/Authority.java
@@ -1,0 +1,7 @@
+package com.jjikmuk.sikdorak.user.user.domain;
+
+public enum Authority {
+
+	USER, ANONYMOUS
+
+}
