+++ be/src/test/java/com/jjikmuk/sikdorak/integration/user/auth/OAuthArgumentResolverIntegrationTest.java
@@ -3,9 +3,9 @@ package com.jjikmuk.sikdorak.integration.user.auth;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.jjikmuk.sikdorak.integration.InitIntegrationTest;
-import com.jjikmuk.sikdorak.user.auth.controller.Authority;
 import com.jjikmuk.sikdorak.user.auth.controller.LoginUser;
 import com.jjikmuk.sikdorak.user.auth.controller.OAuthUserArgumentResolver;
+import com.jjikmuk.sikdorak.user.user.domain.Authority;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
